Switched to using the new status everywhere
