+++ covalent/__init__.py
@@ -29,7 +29,7 @@ from ._dispatcher_plugins import local_dispatch_sync as dispatch_sync  # nopycln
 from ._file_transfer import strategies as fs_strategies  # nopycln: import
 from ._results_manager.results_manager import cancel, get_result, sync  # nopycln: import
 from ._shared_files.config import get_config, reload_config, set_config  # nopycln: import
-from ._shared_files.util_classes import RESULT_STATUS as status  # nopycln: import
+from ._shared_files.statuses import RESULT_STATUS as status  # nopycln: import
 from ._workflow import (  # nopycln: import
     DepsBash,
     DepsCall,
